swagiaal: DwarfDie.getOffset() now returns long instead of int.
frysk-core/frysk/debuginfo/ChangeLog
+2007-12-07  Sami Wagiaalla  <swagiaal@redhat.com>
+
+       DwarfDie.getOffset() now returns long.
+       * TypeEntry.java: Updated.
+

frysk-sys/lib/dwfl/ChangeLog
+2007-12-07  Sami Wagiaalla  <swagiaal@redhat.com>
+
+       * DwarfDie.java (getOffset): Now returns long.
+       * cni/DwarfDie.cxx: Updated.
+

+++ frysk-core/frysk/debuginfo/TypeEntry.java
@@ -342,16 +342,16 @@ public class TypeEntry
 	else
 	    type = typeDie;
 
-	Type mappedType = (Type)dieHash.get(new Integer(type.getOffset()));
+	Type mappedType = (Type)dieHash.get(new Long(type.getOffset()));
 	if (mappedType != null)
 	    return mappedType;
-	else if (dieHash.containsKey(new Integer(type.getOffset()))) {
+	else if (dieHash.containsKey(new Long(type.getOffset()))) {
 	    // ??? will this always be a pointer to ourselves?
 	    // Instead of VoidType, we need a way to reference ourselves
 	    return new PointerType("", byteorder, getByteSize(type),  
 		    new VoidType());
 	}
-	dieHash.put(new Integer(type.getOffset()), null);
+	dieHash.put(new Long(type.getOffset()), null);
 	Type returnType = null;
 	
 	switch (type.getTag().hashCode()) {
@@ -434,7 +434,7 @@ public class TypeEntry
 	}
 
 	if (returnType != null) {
-	    dieHash.put(new Integer(type.getOffset()), returnType);
+	    dieHash.put(new Long(type.getOffset()), returnType);
 	    return returnType;
 	}
 	else
